+++ sdk/identity/azure-identity-broker/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### Other Changes
 
+#### Dependency Updates
+
+- Upgraded `msal4j` from `1.23.1` to version `1.26.0`.
+
 ## 1.1.22 (2026-08-24)
 
 ### Other Changes